Combine tests to one file
